core: Remove last references to AbstractServerImplBuilder
